Changed poller logic to allow stopping poller.  Fixed bugs in asynInt32Array interface.
Changed poller to use epicsEventWaitWithTimeout if poll_delay > 0, or epicsEventWait if poll_delay <= 0.
Thus, the poller can now be disabled by setting poll_delay to 0.
The asynInt32 interface with drvUser=MODDBUS_READ previously called doModbusIO.
It now just sends an epicsEvent to the poller thread to wake it up.
Changing poll_time will also wake up the poller thread, in case the previous value of poll_time was 0 or large.
Fixed problems with the asynInt32Array interface.
  - It was not working if the data type was anything other than UINT16 or INT16.
  - It did not use the ADDR field to set an offset where the array should begin,
    it always started at the first register address for that driver.
It now uses the offset, and uses the data type of the first register to be read when converting from the register
data into the array.